proco smooth for imagenet training

+++ train/train_contrastive.py
@@ -718,7 +718,7 @@ def train_imagenet_inatural(rank, world_size, config, console):
     elif config.training_contrastive.loss == 'procos':
         criterion_ce = LogitAdjust(cls_num_list, device=device)
         criterion_scl = ProCoSLoss(contrast_dim=config.training_contrastive.feat_dim,
-                                   class_frequencies=cls_num_list,
+                                   class_frequencies=torch.FloatTensor(cls_num_list),
                                    temperature=config.training_contrastive.temp,
                                    num_classes=config.sampling.num_class,
                                    device=device)
